pyproject.toml
[project]
name = "smartsnack"
version = "0.18.0"
requires-python = ">=3.10"
dependencies = [
    "flask==3.1.0",
    "gunicorn==23.0.0",
    "pyopenssl==24.3.0",
    "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
e2e = [
    "pytest>=8.0",
    "pytest-playwright>=0.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["blueprints", "services", "app", "config", "db", "helpers", "translations"]
omit = ["tests/*", "tests/e2e/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.",
]
